# openapi.yml
openapi: 3.0.3
info:
  title: Office Commute API
  description: |
    사내 출퇴근/근태/연차/조직 관리 시스템 API.
    모든 응답 오류는 `ErrorResult { code, message }` 또는
    `ValidationErrorResult { code, message, fieldErrorResults }` 형식을 따른다.
  version: 0.0.1

servers:
  - url: http://localhost:8080
    description: 로컬 개발

tags:
  - name: Auth
    description: 로그인/로그아웃 (세션 기반)
  - name: Team
    description: 팀 등록/조회
  - name: Employee
    description: 직원 등록/조회/팀 변경
  - name: Commute
    description: 출퇴근 시간 기록 및 월별 근무 시간 조회
  - name: AnnualLeave
    description: 연차 신청 및 잔여 연차 조회
  - name: OverTime
    description: 초과근무 조회/공휴일 동기화/Excel 리포트

# ----------------------------------------------------------------------
# 권한 모델 (참고)
# - 세션 기반 인증. `JSESSIONID` 쿠키로 식별.
# - `ManagerOnly` 표시된 엔드포인트는 currentRole=MANAGER 만 접근.
# - 비로그인: 401 UNAUTHORIZED. 로그인했으나 권한 부족: 403 FORBIDDEN.
# ----------------------------------------------------------------------
security:
  - SessionAuth: []

paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      description: |
        성공 시 새 HTTP 세션을 생성하고 `currentEmployeeId`, `currentRole`을 저장한다.
        기존 세션은 무효화된다(session fixation 방어).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: 로그인 성공. 응답 본문 없음. Set-Cookie로 JSESSIONID 발급. }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: 이메일 미존재 또는 비밀번호 불일치
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }

  /api/auth/me:
    get:
      tags: [Auth]
      summary: 현재 로그인한 사용자 조회
      description: |
        세션에 저장된 `currentEmployeeId`로 현재 사용자를 조회한다.
        SPA 부팅 시 로그인 상태 복원 및 역할 기반 화면 분기에 사용한다.
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUserResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /api/auth/logout:
    post:
      tags: [Auth]
      summary: 로그아웃
      description: 현재 세션을 무효화한다. 세션이 없어도 200으로 처리한다.
      security: []
      responses:
        '200': { description: 로그아웃 처리됨. 응답 본문 없음. }

  /api/team:
    post:
      tags: [Team]
      summary: 팀 등록 (Manager Only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeamRegisterRequest' }
      responses:
        '201':
          description: 등록 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeamRegisterResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: 동일 이름의 팀이 이미 존재 (TEAM_ALREADY_EXISTS)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
    get:
      tags: [Team]
      summary: 팀 전체 조회
      description: |
        `memberCount`는 저장된 카운터가 아니라 `employee.team_id`의 COUNT 결과를 응답 시점에 합성한 값이다.
      responses:
        '200':
          description: 팀 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TeamFindResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /api/employee:
    post:
      tags: [Employee]
      summary: 직원 등록 (Manager Only)
      description: |
        `teamId`는 옵셔널. 없으면 미배정 직원으로 저장.
        `password`는 BCrypt로 인코딩되어 저장된다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmployeeSaveRequest' }
      responses:
        '201':
          description: 등록 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeRegisterResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: teamId가 존재하지 않음 (TEAM_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
        '409':
          description: employeeCode 또는 email 중복 (EMPLOYEE_ALREADY_EXISTS)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
    get:
      tags: [Employee]
      summary: 직원 전체 조회 (Manager Only)
      responses:
        '200':
          description: 직원 목록. 미배정 직원은 `teamId`/`teamName`이 null.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EmployeeFindResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  /api/employee/{employeeId}/team:
    put:
      tags: [Employee]
      summary: 직원 팀 변경 (Manager Only)
      description: |
        `teamId`가 null이면 미배정 상태로 변경.
      parameters:
        - in: path
          name: employeeId
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmployeeChangeTeamRequest' }
      responses:
        '200': { description: 변경됨 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: |
            employeeId 또는 teamId 미존재.
            코드: `EMPLOYEE_NOT_FOUND` 또는 `TEAM_NOT_FOUND`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }

  /api/commute:
    post:
      tags: [Commute]
      summary: 출근 등록
      description: 현재 로그인 직원의 출근 시각을 서버 시각으로 기록.
      responses:
        '200': { description: 기록됨. 응답 본문 없음. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: 세션의 직원이 존재하지 않음 (EMPLOYEE_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
        '409':
          description: |
            상태/중복 충돌.
            - PREVIOUS_COMMUTE_NOT_ENDED: 다른 날 근무가 종료되지 않음
            - DUPLICATE_WORK: 같은 날 이미 출근 기록 존재
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
        '500':
          description: 예상치 못한 데이터 제약 위반 (DATA_INTEGRITY_ERROR)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
    put:
      tags: [Commute]
      summary: 퇴근 등록
      description: 현재 로그인 직원의 퇴근 시각을 서버 시각으로 기록.
      responses:
        '200': { description: 기록됨. 응답 본문 없음. }
        '400':
          description: |
            제약 위반.
            - COMMUTE_NOT_STARTED: 진행 중인 출근 기록이 없음
            - INVALID_COMMUTE_RANGE: 퇴근 시간이 출근 시간보다 이름
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: 세션의 직원이 존재하지 않음 (EMPLOYEE_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
        '409':
          description: 이미 퇴근 처리됨 (COMMUTE_ALREADY_ENDED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
    get:
      tags: [Commute]
      summary: 월별 근무 시간 조회
      description: 현재 로그인 직원의 특정 년/월 근무 시간을 일자별 + 합계로 반환.
      parameters:
        - in: query
          name: yearMonth
          required: true
          schema: { type: string, example: "2024-08" }
          description: ISO yyyy-MM
      responses:
        '200':
          description: 월별 근무 시간
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkDurationPerDateResponse' }
        '400':
          description: |
            요청 파라미터 오류.
            - MISSING_PARAMETER: yearMonth 누락
            - INVALID_PARAMETER: yearMonth 형식 오류 (yyyy-MM 아님)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /api/annual-leave:
    post:
      tags: [AnnualLeave]
      summary: 연차 신청 (현재 로그인 직원)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnnualLeaveEnrollRequest' }
      responses:
        '200':
          description: 등록된 연차 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AnnualLeaveEnrollmentResponse' }
        '400':
          description: |
            요청 검증 실패.
            - VALIDATION_ERROR: 입력 형식 오류
            - ANNUAL_LEAVE_CRITERIA_NOT_MET: 팀 연차 등록 기준 미충족
            - ANNUAL_LEAVE_PAST_DATE: 과거 일자 신청
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: 세션의 직원이 존재하지 않음 (EMPLOYEE_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
        '409':
          description: |
            상태/중복 충돌.
            - ANNUAL_LEAVE_DUPLICATE: 동일 일자 중복 신청
            - EMPLOYEE_WITHOUT_TEAM: 팀 미배정 직원의 신청
            - DUPLICATE_WORK: 신청 일자에 이미 출근 기록 존재
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }
    get:
      tags: [AnnualLeave]
      summary: 잔여 연차 조회 (현재 로그인 직원)
      responses:
        '200':
          description: 잔여 연차
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnnualLeaveGetRemainingResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /api/overtime:
    get:
      tags: [OverTime]
      summary: 월별 초과근무 계산 (Manager Only)
      parameters:
        - $ref: '#/components/parameters/YearMonthQuery'
      responses:
        '200':
          description: 직원별 연장·휴일근로 분 (근로기준법 기준 — 일 8시간·주 40시간 초과, 휴일근로 별도 트랙)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OverTimeCalculateResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503':
          description: 공휴일 데이터 미가용 (HOLIDAY_DATA_UNAVAILABLE)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }

  /api/overtime/report/excel:
    get:
      tags: [OverTime]
      summary: 월별 초과근무 Excel 리포트 다운로드 (Manager Only)
      parameters:
        - $ref: '#/components/parameters/YearMonthQuery'
      responses:
        '200':
          description: |
            xlsx 바이너리.
            Content-Disposition: `attachment; filename*=UTF-8''YYYY년M월_초과근무보고서.xlsx`
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503':
          description: 공휴일 데이터 미가용 (HOLIDAY_DATA_UNAVAILABLE)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResult' }

components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID

  parameters:
    YearMonthQuery:
      in: query
      name: yearMonth
      required: true
      schema: { type: string, example: "2024-08" }
      description: ISO yyyy-MM

  responses:
    BadRequest:
      description: 잘못된 요청 (BAD_REQUEST / INVALID_JSON 등)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResult' }
    ValidationError:
      description: 입력값 검증 실패 (VALIDATION_ERROR)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResult' }
    Unauthorized:
      description: 비로그인 (UNAUTHORIZED)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResult' }
    Forbidden:
      description: 권한 부족 (FORBIDDEN)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResult' }

  schemas:
    # ---- Common ----
    ErrorResult:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            에러 코드. 예시:
            - BAD_REQUEST, INVALID_JSON, INVALID_PARAMETER, MISSING_PARAMETER
            - UNAUTHORIZED, FORBIDDEN
            - TEAM_ALREADY_EXISTS, EMPLOYEE_ALREADY_EXISTS
            - TEAM_NOT_FOUND, EMPLOYEE_NOT_FOUND
            - DUPLICATE_WORK, DATA_INTEGRITY_ERROR
            - HOLIDAY_DATA_UNAVAILABLE
            - INTERNAL_SERVER_ERROR
        message: { type: string }
    FieldErrorResult:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    ValidationErrorResult:
      type: object
      required: [code, message, fieldErrorResults]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: 입력값이 올바르지 않습니다 }
        fieldErrorResults:
          type: array
          items: { $ref: '#/components/schemas/FieldErrorResult' }

    Role:
      type: string
      enum: [MANAGER, MEMBER]

    # ---- Auth ----
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }

    CurrentUserResponse:
      type: object
      required: [employeeId, name, email, role]
      properties:
        employeeId: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        teamId: { type: integer, format: int64, nullable: true }
        teamName: { type: string, nullable: true }

    # ---- Team ----
    TeamRegisterRequest:
      type: object
      required: [teamName]
      properties:
        teamName: { type: string }
        managerName:
          type: string
          nullable: true
          description: 옵셔널. 빈 문자열은 null로 정규화됨.
        annualLeaveCriteria:
          type: integer
          format: int32
          minimum: 0
          nullable: true
          default: 0
          description: 팀 연차 등록 기준. null 또는 생략 시 0으로 처리됨.
    TeamRegisterResponse:
      type: object
      required: [teamId]
      properties:
        teamId: { type: integer, format: int64 }
    TeamFindResponse:
      type: object
      required: [teamId, name, annualLeaveCriteria, memberCount]
      properties:
        teamId: { type: integer, format: int64 }
        name: { type: string }
        managerName: { type: string, nullable: true }
        annualLeaveCriteria:
          type: integer
          format: int32
          minimum: 0
        memberCount:
          type: integer
          format: int64
          description: employee.team_id COUNT 파생값.

    # ---- Employee ----
    EmployeeSaveRequest:
      type: object
      required: [name, role, birthday, workStartDate, employeeCode, email, password]
      properties:
        name: { type: string, description: 직원 이름 }
        role: { $ref: '#/components/schemas/Role' }
        birthday:
          type: string
          format: date
          description: 과거 날짜
        workStartDate:
          type: string
          format: date
          description: 오늘 또는 과거 날짜
        employeeCode:
          type: string
          pattern: '^[A-Z0-9]{6,10}$'
          description: 사번 (대문자/숫자 6-10자)
        email: { type: string, format: email }
        password:
          type: string
          minLength: 8
          description: 평문. 서버에서 BCrypt로 인코딩.
        teamId:
          type: integer
          format: int64
          nullable: true
          description: 옵셔널. 없으면 미배정.
        timezone:
          type: string
          nullable: true
          example: 'Asia/Seoul'
          description: |
            직원의 기본 시간대 (IANA ZoneId). 옵셔널 — null이거나 빈 문자열이면 서버 기본값 `Asia/Seoul` 적용.
            유효하지 않은 ZoneId면 400 `VALIDATION_ERROR`.
            출퇴근 등록 시 이 값을 기준으로 `workDate`가 계산되며, 각 출근 레코드에 snapshot으로 저장된다.
    EmployeeRegisterResponse:
      type: object
      required: [employeeId]
      properties:
        employeeId: { type: integer, format: int64 }
    EmployeeFindResponse:
      type: object
      required: [employeeId, name, employeeCode, email, role, birthday, workStartDate, timezone]
      properties:
        employeeId: { type: integer, format: int64 }
        teamId: { type: integer, format: int64, nullable: true }
        teamName: { type: string, nullable: true }
        name: { type: string }
        employeeCode: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        birthday: { type: string, format: date }
        workStartDate: { type: string, format: date }
        timezone: { type: string, example: 'Asia/Seoul', description: 'IANA ZoneId — 직원의 기본 시간대.' }
    EmployeeChangeTeamRequest:
      type: object
      properties:
        teamId:
          type: integer
          format: int64
          nullable: true
          description: null이면 미배정 상태로 변경.

    # ---- AnnualLeave ----
    AnnualLeaveEnrollRequest:
      type: object
      required: [wantedDates]
      properties:
        wantedDates:
          type: array
          minItems: 1
          items: { type: string, format: date }
          description: 미래 날짜 목록.
    AnnualLeaveEnrollmentResponse:
      type: object
      required: [annualLeaveId, enrolledDate]
      properties:
        annualLeaveId: { type: integer, format: int64 }
        enrolledDate: { type: string, format: date }
    AnnualLeave:
      type: object
      required: [id, employeeId, wantedDate]
      properties:
        id: { type: integer, format: int64 }
        employeeId: { type: integer, format: int64 }
        wantedDate: { type: string, format: date }
    AnnualLeaveGetRemainingResponse:
      type: object
      required: [employeeId, remainingLeaves]
      properties:
        employeeId: { type: integer, format: int64 }
        remainingLeaves:
          type: array
          items: { $ref: '#/components/schemas/AnnualLeave' }

    # ---- Commute ----
    CommuteDetail:
      type: object
      required: [date, usingDayOff]
      properties:
        date: { type: string, format: date }
        usingDayOff: { type: boolean }
    WorkDurationPerDateResponse:
      type: object
      required: [details, sumWorkingMinutes]
      properties:
        details:
          type: array
          items: { $ref: '#/components/schemas/CommuteDetail' }
        sumWorkingMinutes: { type: integer, format: int64 }

    # ---- OverTime ----
    OverTimeCalculateResponse:
      type: object
      required: [id, employeeCode, name, teamName, overTimeMinutes, holidayWithin8HoursMinutes, holidayExceeding8HoursMinutes]
      properties:
        id: { type: integer, format: int64 }
        employeeCode: { type: string }
        name: { type: string }
        teamName: { type: string, nullable: true }
        overTimeMinutes:
          type: integer
          format: int64
          description: 연장근로 분 — 1일 8시간 초과분 + 1주(월~일) 40시간 초과 잔여분, 가산 1.5배
        holidayWithin8HoursMinutes:
          type: integer
          format: int64
          description: 휴일근로(일요일·공휴일) 중 1일 8시간 이내 분, 가산 1.5배
        holidayExceeding8HoursMinutes:
          type: integer
          format: int64
          description: 휴일근로 중 1일 8시간 초과 분, 가산 2.0배